+++ tests/testthat/test-plot.R
@@ -203,7 +203,7 @@ test_that("plots, relation, multiple CI", {
   set.seed(123)
   vdiffr::expect_doppelganger(
     "plot-relation-multiple-ci-1",
-    plot(em)
+    plot(em, show_data = TRUE)
   )
 })
 
@@ -216,13 +216,13 @@ test_that("plots, estimate_means works with Poisson", {
   set.seed(123)
   vdiffr::expect_doppelganger(
     "plot-means-poisson-1",
-    plot(x)
+    plot(x, show_data = TRUE)
   )
   x <- estimate_means(dat_glm, "fa", backend = "marginaleffects")
   set.seed(123)
   vdiffr::expect_doppelganger(
     "plot-means-poisson-2",
-    plot(x)
+    plot(x, show_data = TRUE)
   )
 })
 
@@ -362,13 +362,13 @@ test_that("plots, automatically join dots", {
   set.seed(123)
   vdiffr::expect_doppelganger(
     "plot-join-dots-1",
-    plot(out)
+    plot(out, show_data = TRUE)
   )
   out <- estimate_expectation(m, by = c("Species", "Petal.Width"), preserve_range = FALSE)
   set.seed(123)
   vdiffr::expect_doppelganger(
     "plot-join-dots-2",
-    plot(out)
+    plot(out, show_data = TRUE)
   )
 })
 
@@ -380,7 +380,7 @@ test_that("plots, logistic regression", {
   set.seed(123)
   vdiffr::expect_doppelganger(
     "plot-glm-logistic-1",
-    plot(out)
+    plot(out, show_data = TRUE)
   )
 })